Add SearchContract.
