+++ mastodon4j/src/main/java/com/sys1yagi/mastodon4j/api/method/Search.kt
@@ -4,18 +4,19 @@ import com.sys1yagi.mastodon4j.MastodonClient
 import com.sys1yagi.mastodon4j.Parameter
 import com.sys1yagi.mastodon4j.api.entity.Results
 import com.sys1yagi.mastodon4j.api.exception.Mastodon4jRequestException
+import com.sys1yagi.mastodon4j.api.method.contract.SearchContract
 
 /**
  * See more https://github.com/tootsuite/documentation/blob/master/Using-the-API/API.md#search
  */
-class Search(val client: MastodonClient) {
+class Search(val client: MastodonClient) : SearchContract.Public, SearchContract.AuthRequired {
 
     /**
      * GET /api/v1/search
      * q: The search query
      * resolve: Whether to resolve non-local accounts
      */
-    fun getSearch(query: String, resolve: Boolean = true): Results {
+    override fun getSearch(query: String, resolve: Boolean): Results {
         val response = client.get(
                 "search",
                 Parameter()
