+++ mastodon4j/src/test/assets/search.json
@@ -1,0 +1,114 @@
+{
+  "accounts": [
+    {
+      "id": 784,
+      "username": "test",
+      "acct": "test",
+      "display_name": "たこやき",
+      "locked": false,
+      "created_at": "2017-04-14T19:10:44.429Z",
+      "followers_count": 22,
+      "following_count": 4,
+      "statuses_count": 5,
+      "note": "",
+      "url": "https://mstdn.jp/@test",
+      "avatar": "/avatars/original/missing.png",
+      "avatar_static": "/avatars/original/missing.png",
+      "header": "/headers/original/missing.png",
+      "header_static": "/headers/original/missing.png"
+    },
+    {
+      "id": 34131,
+      "username": "tester98",
+      "acct": "tester98",
+      "display_name": "tester98",
+      "locked": false,
+      "created_at": "2017-04-15T10:11:09.511Z",
+      "followers_count": 20,
+      "following_count": 1,
+      "statuses_count": 4,
+      "note": "ゲームと手芸とうさぎ好き\r\nTwitterと同じアカ名です",
+      "url": "https://mstdn.jp/@tester98",
+      "avatar": "https://media.mstdn.jp/images/accounts/avatars/000/034/131/original/3466032df9e2cc6d.png?1492251423",
+      "avatar_static": "https://media.mstdn.jp/images/accounts/avatars/000/034/131/original/3466032df9e2cc6d.png?1492251423",
+      "header": "/headers/original/missing.png",
+      "header_static": "/headers/original/missing.png"
+    },
+    {
+      "id": 143489,
+      "username": "testid",
+      "acct": "testid@pawoo.net",
+      "display_name": "testid(pawoo) ☑",
+      "locked": false,
+      "created_at": "2017-04-21T23:15:11.951Z",
+      "followers_count": 2,
+      "following_count": 2,
+      "statuses_count": 27,
+      "note": "ごく稀に絵を描いたりしています(現在休止状態)。少しマニアックな話や趣味全開な内容等の発言中心の方向。\n\nTwitter ：@tes_tid\nniconico ： <a href=\"https://friends.nico/@testid\" rel=\"nofollow noopener\"><span class=\"invisible\">https://</span><span class=\"\">friends.nico/@testid</span><span class=\"invisible\"></span></a>\n鍵付 ： <a href=\"https://mstdn.jp/@testid\" rel=\"nofollow noopener\"><span class=\"invisible\">https://</span><span class=\"\">mstdn.jp/@testid</span><span class=\"invisible\"></span></a>",
+      "url": "https://pawoo.net/@testid",
+      "avatar": "https://media.mstdn.jp/images/accounts/avatars/000/143/489/original/85b577bf411017ee.JPG?1492816512",
+      "avatar_static": "https://media.mstdn.jp/images/accounts/avatars/000/143/489/original/85b577bf411017ee.JPG?1492816512",
+      "header": "https://media.mstdn.jp/images/accounts/headers/000/143/489/original/23cb6971b013a71d.png?1492816512",
+      "header_static": "https://media.mstdn.jp/images/accounts/headers/000/143/489/original/23cb6971b013a71d.png?1492816512"
+    },
+    {
+      "id": 153028,
+      "username": "testid",
+      "acct": "testid@nicodon.jp",
+      "display_name": "testid(隠れ家) :white_check_mark:",
+      "locked": true,
+      "created_at": "2017-04-22T23:42:02.317Z",
+      "followers_count": 1,
+      "following_count": 1,
+      "statuses_count": 10,
+      "note": "隠れ家的なアカウントです（こっそり）\nなので基本非公開とさせて頂きます（フォロー＆フォロバはきまぐれ）\n\nTwitter ：@tes_tid\n<a href=\"https://friends.nico/@testid\" rel=\"nofollow noopener\"><span class=\"invisible\">https://</span><span class=\"\">friends.nico/@testid</span><span class=\"invisible\"></span></a>\n<a href=\"https://pawoo.net/@testid\" rel=\"nofollow noopener noopener\"><span class=\"invisible\">https://</span><span class=\"\">pawoo.net/@testid</span><span class=\"invisible\"></span></a>\n<a href=\"https://mstdn.jp/@testid\" rel=\"nofollow noopener noopener noopener\"><span class=\"invisible\">https://</span><span class=\"\">mstdn.jp/@testid</span><span class=\"invisible\"></span></a>",
+      "url": "https://nicodon.jp/@testid",
+      "avatar": "https://media.mstdn.jp/images/accounts/avatars/000/153/028/original/f6589f7e8c0dec73.png?1492904522",
+      "avatar_static": "https://media.mstdn.jp/images/accounts/avatars/000/153/028/original/f6589f7e8c0dec73.png?1492904522",
+      "header": "https://media.mstdn.jp/images/accounts/headers/000/153/028/original/61a6963ccfc95423.jpg?1492938386",
+      "header_static": "https://media.mstdn.jp/images/accounts/headers/000/153/028/original/61a6963ccfc95423.jpg?1492938386"
+    },
+    {
+      "id": 116733,
+      "username": "testid",
+      "acct": "testid@friends.nico",
+      "display_name": "testid :white_check_mark:",
+      "locked": false,
+      "created_at": "2017-04-20T06:13:12.952Z",
+      "followers_count": 11,
+      "following_count": 10,
+      "statuses_count": 161,
+      "note": "とりあえずニコ生（鏡音系動画上映）、ニコ百（記事編集）をぼちぼちやってたりしてます。ニコニコやアニメ・ゲーム・実況や柔らかめ中心。:nicoru:\n\nTwitter ：@tes_tid\nPawoo ： <a href=\"https://pawoo.net/@testid\" rel=\"nofollow noopener\"><span class=\"invisible\">https://</span><span class=\"\">pawoo.net/@testid</span><span class=\"invisible\"></span></a>\n鍵付 ： <a href=\"https://mstdn.jp/@testid\" rel=\"nofollow noopener noopener\"><span class=\"invisible\">https://</span><span class=\"\">mstdn.jp/@testid</span><span class=\"invisible\"></span></a>",
+      "url": "https://friends.nico/@testid",
+      "avatar": "https://media.mstdn.jp/images/accounts/avatars/000/116/733/original/1023dfb241aad828.png?1492668793",
+      "avatar_static": "https://media.mstdn.jp/images/accounts/avatars/000/116/733/original/1023dfb241aad828.png?1492668793",
+      "header": "https://media.mstdn.jp/images/accounts/headers/000/116/733/original/f940b3f7886d6d73.jpg?1492668793",
+      "header_static": "https://media.mstdn.jp/images/accounts/headers/000/116/733/original/f940b3f7886d6d73.jpg?1492668793"
+    },
+    {
+      "id": 114430,
+      "username": "test",
+      "acct": "test@vocalodon.net",
+      "display_name": "test",
+      "locked": false,
+      "created_at": "2017-04-20T00:17:10.376Z",
+      "followers_count": 0,
+      "following_count": 0,
+      "statuses_count": 1,
+      "note": "テストです。",
+      "url": "https://vocalodon.net/@test",
+      "avatar": "https://media.mstdn.jp/images/accounts/avatars/000/114/430/original/missing.png?1492647430",
+      "avatar_static": "https://media.mstdn.jp/images/accounts/avatars/000/114/430/original/missing.png?1492647430",
+      "header": "https://media.mstdn.jp/images/accounts/headers/000/114/430/original/missing.png?1492647430",
+      "header_static": "https://media.mstdn.jp/images/accounts/headers/000/114/430/original/missing.png?1492647430"
+    }
+  ],
+  "statuses": [],
+  "hashtags": [
+    "testestest",
+    "testset",
+    "test918492",
+    "test7144",
+    "test"
+  ]
+}
